productcontroller
ss

+++ app/Http/Controllers/ProductController.php
@@ -87,11 +87,16 @@ class ProductController extends Controller
             return response()->json(['error' => 'Product not found'], 404);
         }
 
-        $product->delete();
+        if ($product->count > 1) {
+            $product->decrement('count');
+        } else {
+            $product->delete();
+        }
 
         return response()->json(['message' => 'Product deleted successfully']);
     }
-  
+
+
       public function setCategory(Request $request, $productID){
         // Validate the request data if needed
         $request->validate([
